feat: Separate the user document from the user model.
