+++ source/library/models/user.ts
@@ -1,47 +1,24 @@
 import type { Locale, LocalisationLanguage } from "logos:constants/languages/localisation";
 import type { Client } from "logos/client";
+import type { GameType } from "logos/models/documents/guild-statistics/latest";
 import { type IdentifierData, Model } from "logos/models/model";
 import type { DatabaseStore } from "logos/stores/database";
-import type { GameType } from "logos/models/documents/guild-statistics/latest";
-
-interface Account {
-	/**
-	 * User's preferred localisation language.
-	 *
-	 * @since v3.5.0
-	 */
-	language?: LocalisationLanguage;
-
-	/** IDs of servers the user's entry request has been accepted on. */
-	authorisedOn?: string[];
-
-	/** IDs of servers the user's entry request has been rejected on. */
-	rejectedOn?: string[];
-}
-
-interface GameScores {
-	totalScore: number;
-	sessionCount: number;
-}
+import type { GameScores, UserDocument } from "logos/models/documents/user/latest";
 
 type AuthorisationStatus = "authorised" | "rejected";
 
-type CreateUserOptions = {
-	createdAt?: number;
-	account?: Account;
-	scores?: Partial<Record<Locale, Partial<Record<GameType, GameScores>>>>;
-} & IdentifierData<User>;
+type CreateUserOptions = Partial<UserDocument> & IdentifierData<User>;
+
+interface User extends UserDocument {}
 
 class User extends Model<{ collection: "Users"; idParts: ["userId"] }> {
 	static readonly #initialScores: GameScores = { totalScore: 0, sessionCount: 1 };
 
+	readonly createdAt: number;
+
 	get userId(): string {
 		return this.idParts[0];
 	}
-
-	readonly createdAt: number;
-	account?: Account;
-	scores?: Partial<Record<Locale, Partial<Record<GameType, GameScores>>>>;
 
 	get preferredLanguage(): LocalisationLanguage | undefined {
 		return this.account?.language;
